fix(desktop): stop every runtime pair for an agent
